Address code review: consolidate panic hook, shared rendering, terminal size
- Move ratatui panic hook to main.rs (single install instead of per-screen)
- Extract render_particles() and render_logo() into effects.rs, removing
  duplicated rendering code from splash.rs and onboarding.rs
- Move LOGO constant to effects.rs for shared access
- Use crossterm::terminal::size() for particle pre-seeding instead of
  hardcoded 80x24
- Restore literal unicode chars in PARTICLE_CHARS for readability
- Remove redundant drop(terminal) in splash.rs
- Update README.md with splash screen mention
